+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   - fix wrong call in `MetricBase`
   - change default `ncols` of `pbar` into 100
   - add `dump_middle=True` argument to `.eval()` in `SimpleMetricTask`
+  - change logic in `SimpleTask._eval_during_train()` to make sure ckpt dumping
+  - dump model at the beginning of `_eval_during_train` if `save_every_ckpt` in case of any unexpected crashes
 
 - v0.1.7
   - add `SimpleMetricTask` for training and evaluation with metric instances