+++ CHANGELOG.md
@@ -12,6 +12,7 @@
   - add `dump_middle=True` argument to `.eval()` in `SimpleMetricTask`
   - change logic in `SimpleTask._eval_during_train()` to make sure ckpt dumping
   - dump model at the beginning of `_eval_during_train` if `save_every_ckpt` in case of any unexpected crashes
+  - fix mutable arg in `LabelEncoder`: change `initial_dict` into `initial`, default is None
 
 - v0.1.7
   - add `SimpleMetricTask` for training and evaluation with metric instances